+++ G07P3/src/src/seleccion/ISeleccion.java
@@ -9,8 +9,8 @@ import src.utils.Utils;
 public interface ISeleccion {
   public ArrayList<Cromosoma> select(ArrayList<Cromosoma> poblacion, Random rand);
 
-  default public ArrayList<Integer> group(ArrayList<Cromosoma> poblacion){
-		ArrayList<Integer> groupedFitness = new ArrayList<Integer>();
+  default public ArrayList<Double> group(ArrayList<Cromosoma> poblacion){
+		ArrayList<Double> groupedFitness = new ArrayList<Double>();
 		
 		for(Cromosoma i : poblacion){
 			groupedFitness.add(i.getFitness());
@@ -23,7 +23,7 @@ public interface ISeleccion {
 		}
 
 		for(int i = 0; i < groupedFitness.size(); ++i){
-			groupedFitness.set(i, (int)(1.05*max) - groupedFitness.get(i));
+			groupedFitness.set(i, (1.05*max) - groupedFitness.get(i));
 		}
 
 		poblacion = Utils.sortSample(poblacion);
