+++ G07P3/src/src/seleccion/SeleccionUniEstocastica.java
@@ -10,7 +10,7 @@ public class SeleccionUniEstocastica implements ISeleccion{
 
   @Override
   public ArrayList<Cromosoma> select(ArrayList<Cromosoma> poblacion, Random random) {
-	  	ArrayList<Integer> probability = new ArrayList<Integer>();
+	  	ArrayList<Double> probability = new ArrayList<Double>();
 		ArrayList<Cromosoma> selection = new ArrayList<Cromosoma>();
 		double totalFit = 0;
 		int tamPob = poblacion.size();
@@ -21,7 +21,7 @@ public class SeleccionUniEstocastica implements ISeleccion{
 			totalFit += i;
 		
 		for(int i = 0; i < probability.size(); ++i) // Sumamos la probabilidad acumulada, al primero se le suma 0 y al resto se acumula
-			probability.set(i, (int)((probability.get(i) / totalFit) + (i == 0 ? 0 : probability.get(i-1))));
+			probability.set(i, (probability.get(i) / totalFit) + (i == 0 ? 0 : probability.get(i-1)));
 		
 		double distPtr = 1.0 / tamPob; // calculamos la distancias entre las marcas.
 		double rand = random.nextDouble(0, distPtr); // pos de primera marca se obtiene a partir de rand entre [0, distPtr]
